数据库表设计优化：尽量使用int，而非bigint，如果非负则加上unsigned（这样数字扩容会扩大一倍），当然能使用tinyint/smallint/medium_int更好

+++ spring-boot-rbac/spring-boot-rbac-mybatis-plus/src/main/java/com/example/lewjun/domain/SysDeptRole.java
@@ -8,6 +8,6 @@ import lombok.experimental.Accessors;
 @Data
 @Accessors(chain = true)
 public class SysDeptRole extends BaseObj {
-    private Long deptId;
-    private Long roleId;
+    private Integer deptId;
+    private Integer roleId;
 }
